Revert "Merge branch 'master' of https://github.com/Eytins/HotelManagement"
This reverts commit ae9a240e0abc41165bb677b7819337f27f1525ee, reversing
changes made to f7fced49c4f136d5c9aba3df4795a3e0cdd16b78.

+++ src/com/HotelManagement/service/hotelType/HotelTypeService.java
@@ -29,7 +29,7 @@ public interface HotelTypeService {
 
     //更改房型信息
     int modifyRoomType(Integer roomTypeId,
-                       long price,
+                       Integer price,
                        Integer numberOfPerson);
 
 }
